All Code
Every Management System, Billing System is trying to make the automatic, portable and remote control. This paper presents a Smart Energy Metering System for automatic and flexible and billing system. Using the Arduino with Ethernet Shield provide the meter reading system with designing Interrupts as per our requirement so that we could get an accurate Reading. The proposed energy metering system can implement with Open Source Hardware like Arduino and Ethernet Shield to transmit the data, pulses generated by meter, security services, calculated bill over the Internet. In recent years, a wide variety of solutions has been proposed for the data transmission and remote measurement. In this Project, an architecture of a real-time smart energy management with billing feature is presented. It provides several Smart Meters, that continuously monitor connected loads and communicate with Arduino. The main task of Arduino is to catch the pulse of the Meter and then fire a query in the database to store the pulse as well as calculate kWh forwarded from a single smart meter, in order to make this information available on Internet using HTTP server.